refactor(mtp): separate builder components
